+++ wavesplatform/src/main/java/com/wavesplatform/sdk/model/request/node/SetAssetScriptTransaction.kt
@@ -5,7 +5,6 @@ import com.google.common.primitives.Bytes
 import com.google.common.primitives.Longs
 import com.google.gson.annotations.SerializedName
 import com.wavesplatform.sdk.crypto.WavesCrypto
-import com.wavesplatform.sdk.utils.arrayWithSize
 import com.wavesplatform.sdk.utils.scriptBytes
 
 /**
@@ -41,8 +40,8 @@ class SetAssetScriptTransaction(
     override fun toBytes(): ByteArray {
         return try {
             Bytes.concat(
-                byteArrayOf(type.toByte()),
-                byteArrayOf(version.toByte()),
+                byteArrayOf(type),
+                byteArrayOf(version),
                 byteArrayOf(chainId),
                 WavesCrypto.base58decode(senderPublicKey),
                 WavesCrypto.base58decode(assetId),
